--- src/Mandrill/Utilities/MandrillRequestException.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Mandrill.Utilities
-{
-  public class MandrillRequestException : MandrillException
-  {
-    public MandrillRequestException(string message, Exception ex)
-        : base(message, ex)
-    {
-    }
-  }
-}
