+++ src/Mandrill/Utilities/MandrillSerializeRequestException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Mandrill.Utilities
+{
+  public class MandrillSerializeRequestException : MandrillException
+  {
+    public MandrillSerializeRequestException(string message, Exception ex)
+        : base(message, ex)
+    {
+    }
+  }
+}
